Fix excessive cloning when deleting PCVs by ID
The normalization of values to-be-deleted was done too lately,
leading to excessive cloning (basically at each mapping source
evaluation). This is now fixed. The normalization is done when
ODO is computed in the focus context.

This resolves MID-7057. However, it is only an approximate fix
for now. The more thorough analysis of the problem is needed.