repo-sqale: fixed storing of extension refs and their target type
Target type is now always stored, it's never "nullable" in the reference
JSON object. If default target type is provided in schema, it is used
if the Referencable does not provide it.
Storing the type is important for queries using it in WHERE, with no
information there is no correct answer to the query.
If target type is not provided and no default is in schema, exception
is thrown during extension creation (during add) which is good because
such objects would not be gettable anyway (schema error would occur).