MID-6475: Windows service is back, so is running with -jar midpoint.war
The MidPointWarLauncher was reinstated in a simplified form as a first
step before PropertiesLauncher is run.
It contains needed "stop" method for Windows service.
This allowed to initialize loader.path here, which is much cleaner from
the outside perspective, JARs in ${midpoint.home}/lib just work again
even when run as -jar midpoint.war.