repo-sqale: MID-6967 support for assignment/subtype search, EQ only
With one value assignment having subtype with that value is found.
With multiple values in the query, assignment with ANY of the values
is found (this is in line with semantics of multi-value EQ for in memory
filtering).
This is only added for new/future repository, old one would require
subtypes table.
Minor style/cosmetics adjustments around gin index clauses.