Clean up and document activity framework code
1. Now we prefer composition over subclassing. Activity execution
classes are no longer subclasses of an abstract execution class (except
for very special ones). Instead, IterativeActivityExecutionSpecifics
interface is implemented to provide the necessary functionality.

2. SimpleActivityHandler was aligned with these changes. The framework
itself is now sufficiently simple to make such adaptations
and simplifications unnecessary.

3. Tracing- and profiling-related methods in RunningTask interface
were removed. This functionality is related to activities, not to
tasks.

4. A lot of code was reviewed and cleaned up. Key classes and interfaces
of the framework are now documented.

Externally-visible changes:
- OperationExecutionType now contains activity path instead
of obsolete task part URI.
- Profiling/tracing, reporting options and failed objects selector
are now part of activity definition.
- Temporarily removed support for tracing of the pre-processing
stage for live sync and provisioning search-based activities
(like import or reconciliation). This is a result of fixes
in RunningTask interfaces.
- Removed "Last N failures" hack. It is no longer needed: the regular
failure-tracking mechanism should be good enough now.