Fix memleaks caused by log capture in tracing
The log capture in tracing can cause memleaks - along with mess
in the logs gathered - if the operation results are not treated
appropriately. (Which is currently the case in many places in the code.)

So this commit:
1) Simplifies and documents the log capturing mechanism.
2) Tries to reduce the scope of memleaks (should they occur) by
unlinking event sinks with operation result objects.
3) Introduces some sanity checks to the process of switching logging
sinks. These checks help to find problems in operation result handling.
4) When there is a conflict between real and expected log recorder,
we now use expected log recorder to be used, hopefully fixing any
problems stemming from bugs in code.

We also fixed some of specific places where operation result
handling was not correct.