Stop maintaining buckets in worker tasks
It seems that careful and systematic use of modifyObjectDynamically
obsoletes the original design of "borrowing" buckets from coordinator
and storing them in workers.

So this commit:

1) Stops storing buckets in the worker tasks.
2) Migrates all bucket management operations to modifyObjectDynamically
mechanism.

This simplifies the bucketing code, makes it more robust, and reduces
the number of repository operations (e.g. it eliminates all getObject
operations when manipulating the buckets).

Related changes:
- RunningTask has now a reference to (resolved) parent task.