Improve auto-scaling mechanism
1. A task releases all buckets as soon as it exits processing
of a current bucket in a way other than regular completion. This
ensures e.g. that tasks that are suspended in a standard way hold
no buckets. (And this ensures correct processing of all buckets by
remaining workers.)

2. When a node is detected to be down by ClusterManager, all buckets
held by all tasks that were executing on it are automatically released.
The reason is exactly the same as above: to prevent buckets be held
by ghost tasks indefinitely.

3. When reconciling workers, we suspend only the superfluous tasks that
run on nodes that are still up and alive. (It is not sufficient
that they are not down.) This is because we rely on the tasks
themselves to release their buckets. This is not a perfect solution,
though - see MID-7180.