Fix handling of deleted focus in ContextLoader
The context loading in later clockwork clicks (after focus was deleted)
resulted in old, current, and new object being set to the same value
of objectOld. This confused e.g. updated AssignmentProcessor, which
(in 4.4) executes also in cases where focus is being deleted.

The loading was fixed by introducing focusContext.deleted flag. If it
is present, both objectCurrent and objectNew are set to null by the
loader.

Related to MID-4856.