Make deletion-related recon. thresholds work
1. In order to detect exceptions (like threshold violations) in the
"remaining shadows" reconciliation activity, its approach was changed:
instead of relying on built-in ObjectNotFoundHandler we do handling
of the missing objects ourselves. This ensures that any exceptions
are not buried under layers of OperationResult instances, so they
can be processed in ItemProcessingGatekeeper.

2. In order to skip handling ObjectNotFoundExceptions in GET operations,
experimental "Forced exception" error reporting mode
in GetOperationOptions was introduced. This disables
ObjectNotFoundHandler.

3. When throwing ObjectNotFoundException for a get shadow op, we try
to provide an OID. This is to allow the client to distinguish the
unavailability of the requested shadow from unrelated problems.

!!! BEHAVIOR CHANGES in points 4-5 !!!

4. Full synchronization timestamp is NOT set for simulation mode.
Only in execution mode. This also means that when fetching remaining
shadows in recon, we use fullSynchronizationTimestamp in execution mode,
and synchronizationTimestamp in all other modes.

5. Shadows for missing objects are NOT marked as tombstone
(dead/not-existing) in simulation or dry run mode. This is to avoid
loss of this information when it comes to the real execution.

Other changes:

6. Monitoring configuration is now tailorable for sub-activities.

Work in progress. Some tests may fail.