Implement activity execution reporting
An activity can now provide four types of execution reports:
- buckets processed (or analyzed),
- items processed (or skipped),
- ConnId operations executed (within item processing or globally),
- internal operations executed (within item processing).

It is possible to filter individual records that are being written
to these reports - like ConnId ops only for specified resource or only
the ones that took a long time - and for specified reports (ConnId,
internal ops) also enable/disable report chunks for individual items.

Introducing also before/after conditions for writing traces, so it is
now possible e.g. to specify that a trace should be written for each
item that took more than specified time to process, or that failed, etc.

Bucket size analysis mode was added here.

Finally, this commit introduces "item processing condition" that allows
to skip specific items from being processed. (It is expected to be used
for troubleshooting and/or in temporary modifications of activity
behavior.)

All these features are of highly experimental nature.

Related changes:
1. Now it's possible to provide more tracing configurations for
an activity.
2. Each ConnId operation now has a lightweight identifier, to allow
matching ConnId operation report with the detailed log, if needed.
3. ConnId operation reporting is now greatly simplified. It uses
listeners at the level of a Task.

Unrelated change:
 - Fixed ClassicDashboardReportExportActivityExecutionSpecifics#determineExpectedTotal
(+= instead of =+)

Tests are quite rudimentary for now.