Fix activity tree overview structure
The tree overview structure is now aligned with the needs of GUI:
it contains things to be displayed in the task list. So we no longer
have to retrieve the whole task tree to be able to display progress,
errors, and overall status of a task tree.

Properties added:
 - simplified bucket-related progress,
 - per-task entry carrying task state and items-related progress.

Related changes:
 - Fixed working with task-level statistics in activity-based tasks.
 - Renamed legacy progress-related methods in Task API.
 - Removed obsolete items from StatisticsCollectionStrategy.

Fixes MID-7184.