Add fake OperationResultType.incomplete XSD attr
In order to make XML parsers (like those in MidPoint Studio) happy
with XML of repository objects here we introduced fake
"incomplete" attribute in OperationResultType. This is because
tasks are often returned from repo like this:

<task>
  ...
  <result incomplete="true"/>
  ...
</task>

And it looks ugly when all those places are marked in red in Studio.

Related to MID-7023.