Use read-only when fetching objects in model
To speed up processing of objects in clockwork we have introduced
using read-only option when loading objects at a lot of places
in model-common and model-impl.

Limitations:

1) For now, we avoid using read-only option when loading objects for
focus or projection contexts, because these are worked with
during processing. This can be optimized later.

2) In provisioning, shadows are fetched read-write because of the
need to apply definitions right after obtaining them from the repo.

3) At a couple of other places we kept existing (read-write) read mode,
particularly when the perceived benefit was low and the risk of getting
immutability exceptions was high.

Related to MID-7053.