Clean up lens structures before attacking MID-7053
It looks like midPoint spends excessive amount of time
in LensElementContext#getCurrentDelta and
LensFocusContext#getSummaryDelta: 18% of the clockwork time
for 100 assignments, 45% for 500 assignments in TestSystemPerformance.
This has to be changed, e.g. by avoiding frequent (and useless)
recomputation of these deltas.

Before doing that, here is a small code cleanup related
to the clockwork. Main changes are:

1) The clockwork "click" functionality was factored out into
ClockworkClick class.

2) The ContextLoader monolith was broken up into smaller components,
refactored, and slightly documented. (Still not finished, though.)

3) Fixed checks for archetypeRef vs assignments on clockwork start
(didn't account for multiple archetypes).

4) Removed PrismContext-related methods from lens context.

5) Removed deprecated "reconcile affected" feature.

Work in progress.