Make TestLdapMassive work
This commit TEMPORARILY reverts a change that was made in
58315473ee865e950e4f96dd171d919d83ce35d1 (removing resource.namespace).

That change, although correct in itself, made TestLdapMassive.test210
start failing. Perhaps the test itself is broken, and its green status
was only by chance: because of an error in dj namespace declaration
in resource definition.

To do: analyze and maybe fix the test.