Eliminate repeated delta/object recomputations
Each time current or summary delta (including relative and absolute ODO)
were obtained from LensElementContext, full re-creation of these deltas
from primary and secondary deltas occurred. This accounted for large
portion of time spent in the clockwork for some use cases.

This commit partly eliminates these recomputations: the computable
parts of the element state (mainly: current/summary delta, new object)
are cached, with invalidation bound to the changes in "source" data.

Work in progress. The performance can be improved further
by updating current/summary delta "on the fly", along with updates
to secondary deltas. It is also questionable if it is better to
recompute new object when accessed (and invalid), or to use original
approach based on LensElementContext#recompute method, or to update
it "on the fly" along with current/summary delta.

However, even this partial work saved 10-15% clockwork time
for scenarios with 100 assignments, and 30-35% time for 500 assignments.

Unrelated minor changes:

1. ConsolidationProcessor no longer drops primary and secondary deltas
for projections that were deleted.

2. Synchronization intent was moved from LensElementContext
to LensProjectionContext.

Related to MID-7053.