MID-7074: Added extra logic to handle duplicate discovered connector
Connector discovery checks using search if discovered connector is already in database,
but in clustered case - other node can insert it into database between check
and add. If database fails add because of unique constraints, code does not fail
discovery, but rather checks again if matching connector is present.

This prevents multiple additions of same connector, but repository / database combination
needs to support it (different databases needs different definition of index).

As of this patch sqale repository with postgresql is supported.