TestWorkerTasks: Run clusterManagerThread
This test was fast with in-memory database and with new postgresql, so missing updates of node availability
were not visible. Once database is external and uses hibernate local node timeouted, so task rescheduling
did not worked correctly.

Start clusterManager update thread (as in running application) ensured that timestamps in DB are correct
and node is correctly visible during testing.

Signed-off-by: Tony Tkacik <tonydamage@gmail.com>