Review the tasks schema (finally)
This is a (hopefully) final pre-release review of common-tasks-3 schema.

Some types were renamed to achieve a bit of consistency. Obviously
outdated or leftover pieces of schema were removed. Some docs were
added.

Main activity features (and some structures that we use for a long time)
were marked as non-experimental.