Fix assignment-related approvals
Assignment-related policy rules in 4.0 have relied on assignment
being present in plus/zero/minus set. This is wrong, as the
placement in these sets depends on other factors,
e.g. on assignment validity.

This behavior has been fixed in midPoint 4.2 (as part of midPrivacy).

So, here we backport relevant parts of the new code: AssignmentOrigin,
SmartAssignmentCollection, and SmartAssignmentElement, plus a couple
of other ones.

The overall idea is to change policy rules processing ONLY, but do not
touch (fix) placement of assignments in their sets. (This is because fixing
the placement would need many subsequent fixes to be brought from 4.2
to here.)

In addition to that, we have fixed default approval behavior when
no policy rules are present: using "is added" instead of "is in plus set".
See 1e43c8971762b3826d81545b3ca21ff853a05e03 for the same fix
on master.

Other changes:
- Assignment pruning is now based on "being added" instead of on
"in plus set".
- Fixed displaying of policy rule actions in PolicyRuleTypeUtil.

This resolves MID-7317.