Add an extra check for maintenance mode
In this commit we tried to reproduce values being deleted from a focus
when the source resource is in maintenance mode. We were not successful.

Nevertheless, we have implemented an extra check that prevents running
the synchronization of a resource object when the resource is in the
maintenance mode. In addition to that, a check for maintenance mode
was added at the start of ProvisioningServiceImpl#synchronize method.

Also:
- Fixed NPE when displaying specific traces.

Related to MID-7062.