Avoid leaving shadow in reaping state
This is a fix for MID-7386, where an account (that should exist)
was left in "reaping" state leading to problems after shadow refresh
was run. Although we were not able to reproduce the issue fully, here
are some fixes regarding this.

The main problem was that the projector considered shadows in "reaping"
state as having tombstone=true.

To fix this the following changes were done:

- LensProjectionContext#isTombstone was changed to #isGone. The same for
  ResourceShadowDiscriminator. The reason is that this flag covers both
  "tombstone" and "corpse" states. The new name emphasises the fact that
  the resource object is deleted (gone) from the resource.

- Added transient shadowLifecycleState property to ShadowType. This is
  to ensure that the projector knows the exact lifecycle state for
  a shadow being obtained from provisioning. All shadows coming to the
  projector must have this property.

Other changes needed for the fix:

- Shadows in "reaping" state get policy decision = KEEP during
  activation processing.

- Determination of shadow liveness state (needed e.g. for the decision
  about linkRef being org:default or org:related) was fixed.

- Links to shadows in reaping state are treated more carefully
  in the context loader.

Other changes:

- Some shadow lifecycle states were renamed: conception -> conceived,
  gestation -> gestating, life -> live.