Change member operation tasks in GUI to activities
Role member operations (assign, unassign, delete, recompute) that are
implemented in MemberOperationsHelper were using old-style extension-
based task configuration. This was now changed to new activity-based
notation, allowing the tasks be correctly shown and edited in GUI.

Other changes:
- MemberOperationsHelper was refactored in order to make these changes.
- Pure GUI related functionality was factored out to
MemberOperationsGuiHelper.
- "Old" MemberOperationsHelper in web/page tree was converted to be
just a wrapper for the "new" implementation, and renamed to
MemberOperationsHelperOld (to avoid ugly package names in the code).