Migrate DeleteTaskHandler to activity
This commit creates DeletionActivityHandler as a replacement for the
legacy handler code.

It also fixes OperationResult-handling code in tasks/activities:

1. We definitely got rid of TaskRunResult#operationResult property.
So the task handlers can no longer provide their "own" result.
They should use Task#getResult() or create a subresult of it.
Legacy task handlers were adapted to this change.

2. Therefore, TaskRunResult and ActivityRunResult got properties that
keep the exception responsible for the run failure, plus TRR also the
message. These properties (if present) are used to update task's
root operation result.

3. The Task#refresh method no longer updates live operation result
for RunningTask instances. Quite the opposite: it updates in-prism
version of the result from the live one.

Unrelated change:

- Fixed root xsd:element definitions in common-tasks-3.xsd.