Stop de-registering ActivityBasedTaskHandler
It looks like that de-registering ActivityBasedTaskHandler during system
shutdown causes problems if recurring tasks are started at the very
same moment (MID-7328, MID-7331).

The issue has to be investigated more thoroughly, but as an immediate
workaround here we simply disable the code that de-registers
ActivityBasedTaskHandler from the list of handlers. (Note that
in 4.3 and before, the de-registration of task handler did not exist
at all.)

Unrelated changes:
- Not logging bucket completion for LiveSync tasks by default
(eliminates log noise).
- Removed some unneeded features of TaskHandlerRegistry.