Use urandom for generating secrets in midpoint.sh
Init procedure in midpoint.sh script needs to use /dev/urandom as a
source of randomness. It was using /dev/random, but it can block for a
very long time when the system doesn't have enough entropy. This was
sometimes happening when the script was started from a fresh docker
container.