Change recomputation task default to FocusType
Before 4.4 the default type was UserType. During 4.4 development
we have changed it to AssignmentHolderType, but this does not make
much sense. In fact, FocusType is the kind of objects that is generally
"recomputable".

(Moreover, when recomputing the recomputation task itself, an error
occurs: "The task:... refers to itself in assignment/inducement".)