new repo: fix of missing containerType conditions for assignments/indc.
This affects only AbstractRole subtypes, but the problem was that filter
on F_ASSIGNMENT and F_INDUCEMENT were not distinguished properly.
This is fixed now in the mapping, in addition to the owner OID condition
on the ON clause, proper containerType condition is added as well.

(cherry picked from commit 5971d8f4a879f4246cff30472f94e518cefc189e)