Fix "incomplete but having value" issue in tasks
When e.g. refreshing a task, we unintentionally created a task prism
object with result marked as incomplete and having a value at the same
time.

A serialized form of such tasks (e.g. in trace files) was then
not parseable because of "Attempt to store multiple values in
single-valued property" exception.

Resolves MID-7583.