Fix activity statistics log records
1) The import, reconciliation, and live sync activities did not call
parent createReportingCharacteristics() method, so the upstream settings
(namely, creation of run records) was not applied. Therefore,
the wall-clock time and throughput was not measured for import
and reconciliation tasks. (For live sync the run records are not created
intentionally.)

2) When information about overall wall-clock time (obtained from run
records) is missing, no "null" values are logged.

This resolves MID-7585.