Make ID Match integration more user-centric
The correlation is now based on matching midPoint user properties,
instead of shadow attributes as it was before. The result of ID Match
correlation is the owner OID instead of (raw) Reference ID.

Also:

1. Correlation context in cases is simplified. No more correlator-
specific content is there. Options are represented purely as owner
references (with confidence). Shadow is still there, but the correlated
object is primarily represented as focus object.

2. This means that ID Match correlator now uses focus instead
of resource object as a source of data to be sent to the ID Match
service. Note that pre-mappings (i.e. pre-clockwork execution of
real inbound mappings) are only simulated now.

3. ID Match uses a default SOR of "midPoint". SOR IDs are prefixed
by configurable string. The matchgrid ID is now configurable as part
of ID Match service URL.

4. Improved correlation GUI: configurable correlation properties,
ability to distinguish primary and secondary values of these properties,
distinguishing full (green) and partial (yellow) matches.

Work in progress.