notification schema: added new message transport configuration
This is mostly copy/paste of the old one, but name is now an element,
not an attribute - old types were deprecated, but some configuration
details were reused (e.g. mail server config).
New types have the word "Transport" added, but to make this consistent
old CustomTransportConfigurationType was renamed to Legacy*, because
it is already the right name we want to use in the future.