Clean up inbounds processing before pre-mappings
Here we cleaned-up the processing of inbound mappings, in the order
to separate parts bound to the clockwork processing from these that
can be executed before the clockwork is even started.

Work in progress. The code is not cleaned up absolutely. But it is
the maximum that could be done in the given time.

BEHAVIORAL CHANGES (to release notes):

1. Up to now, we have skipped association inbound mappings if there was
no input value. This was not consistent with the handling of other
mappings. This is now changed. So, please make sure your mappings
correctly treat null input values. See resource-dummy-autogreen.xml
for an example.

2. Names of classes and packages that are relevant for logging has
been changed. To log the inbound processing in details, one should
now use "com.evolveum.midpoint.model.impl.lens.projector.focus.inbounds"
(however, this may even change before 4.5 release!)

Unrelated changes:
 - contains changes from dcdc09e31d80b5b2376ff80e6d1e6155c035c29b