Update default case notifier to correlation cases
The notifier was treating all cases as approval ones. This is now
fixed, and the approval terminology will not be shown in correlation
and manual provisioning cases.

Work in progress: the content of notifications is still preliminary.
Moreover, the notifier will be - most probably - replaced by a default
message template.

Related changes:
- Renamed WorkflowEvent to CaseManagementEvent.
- Renamed WorkflowProcessEvent to CaseEvent.
- Introduced wf.caseOid auditing constant (replacing
wf.processInstanceId).

Unrelated changes:
- Renamed NotificationFunctions#getObjectType to getObject.