Loose model-impl/cases-impl dependency
The model-impl was made dependent on cases-impl by autowiring
CaseManager in CorrelationCaseEngineExtension, in order to register
the extension in the manager.

This commit relaxes the dependency by changing the registration
mechanism: it is now done by Spring autowiring, so the extension
can be present even if the CaseManager is not. In such situations,
cases will be created but will not be processed by the case engine.