MID-7484: recipientExpression returns RecipientExpressionResultType
This allows either string or focus values. We need to use focuses more
to enable localization based on user preferences.
Also, sending notification to different transports with different kinds
of addresses is possible this way - string address is not enough for it.
Transport send is now called for each recipient, but message composition
from template is still WIP.