Cleanup ID Match client code
The code was quite "prototypical"; now it should be better - although
not 100% clean. In addition to simplification and cleaning up the
request/response logging and error handling were added. The obsolete
"includeNoneMatchesOption" was removed.

Tests for invalid credentials and invalid address handling were added.

Unrelated but important change:

- SynchronizationType and ObjectSynchronizationType are now containers,
not properties. This solves e.g. the problem that the ID Match service
password was not encrypted in resource object saved in repository.