Re-add $user/$account vars even for non-accounts
By mistake, these variables have been available for correlation
purposes even for resource objects that are not accounts (like
e.g. groups). It has been in this state for a long time. Recently
I have "fixed" this. But it seems that some tests rely on this bug.

So I'm (at least temporarily) re-enabling this behavior, so
$account is available also for groups, orgs, and other non-account
objects.