Fix handling of unknown intents in GUI
As part of cleanup of resource object definitions (object types,
object classes), the method ResourceSchema#findObjectDefinition was
made more picky - it intentionally rejects attempts to find a definition
for kind or intent being unknown. The reason is that the client should
know what it is looking for.

So, in this commit I reviewed calls to this method from GUI, and added
appropriate checks there. However, I am not sure about the expected
behavior, so please REVIEW this commit.

This should resolve MID-7705.