Re-use sequence values during approvals
When a sequence value(s) are used during operation being approved,
they are now stored along with the approval request. After the request
is approved, the same values are applied. Therefore, e.g. for user
creation, the properties of user actually created are the same as
they were in the creation approval request.

This resolves MID-7575.