Add sync transitions for "remaining shadows" recon
The reporting on synchronization situation transitions during the
last reconciliation sub-activity (processing "remaining shadows") was
not implemented: It was somehow present before, but the implementation
changes in 4.3/4.4 made it non-functional, so it was disabled
altogether.

So, it is re-implemented in this commit. In particular:

1. Synchronization of a deleted shadow is represented as a transition
"x->DELETED->DELETED", where x is the original state, e.g. LINKED.

2. Shadows that are fetched but not synchronized - e.g. because they
exist - are recorded with either PROTECTED or with the new
SYNCHRONIZATION_NOT_NEEDED exclusion reason. They are also marked
as skipped, which is a change comparing to previous behavior (they
were marked as successfully processed). This is to bring consistency
between the 2nd and 3rd stage of reconciliation activity.

This resolves MID-7724.

Other changes:
- Fixed reporting of situations of deleted shadows in "clockwork
summary" log entry.
- Re-enabled auditing asserts in TestImportRecon.