Fix working with OperationResult in Handler
OperationResult should be filled-in according to the code call
structure, i.e. called code should use the result of the parent.

The use of Handler interface, passing values to be processed but without
the corresponding OperationResult, is not correct in this respect.

So here are the following changes:

1. Introduced ObjectHandler interface that contains not only the value
but also the OperationResult instance. [TODO: better name?]

2. Used ObjectHandler for iterative processing in reports, instead of
Handler.

3. Renamed ObjectHandler in UCF API to UcfObjectHandler, to avoid naming
collisions.

Related to MID-7737.

(cherry picked from commit 27de9e7a67c6bdb2d762e5322a7cf924fa6bf035)