Reload lens context if there's a focus change
If the focus is changed during lens context loading - presumably
because of the discovery process - we will rot the context and re-load
it. This is to avoid continuing with the obsolete data, because that
can lead to inconsistent results, as demonstrated
by TestVeryAuthoritativeSource (and MID-7725).

This resolves MID-7725.

(cherry picked from commit bb2a191dc5f7353f2ccebc41e036ab22762f3462)