Make TestUnix more tolerant
This test checks a timestamp of account created in OpenDJ. However, the
resource rounds the time to whole seconds. For unknown reasons, it looks
like the safety margin of 1000 millis is not enough, so here it is
increased to 1500.

(cherry picked from commit 621aa69bef2479c43ae449cf1642acc7c06b3c58)