system-init: added support for "midpoint.configFile" system property
Tests could set name for config.xml via constructor, but it was not
possible for production, which is needed for Ninja native repo tests.
If non-default config file is missing, it is NOT extracted as usual.
For consistency "midpointPrintSensitiveValues" was renamed to
"midpoint.printSensitiveValues" (the old one works for a while too).