Fix channel handling in synchronization tasks
The channel URI was not consistently set in synchronization activities
(import, recon, live sync, async update). It was hacked in a complex
way, and still not reliable enough.

This commit sets the task.channel property for such synchronization
activities, hopefully providing channel to all places where it is
needed. Some of the hacks could be thus removed.

This resolves MID-7454.