Stop trying to update low-level stats for AU
There are some threading issues in asynchronous update tasks, e.g. when
AMQP client is used: The thread that handles the connection is different
from the thread that belongs to the async. update task.

We are not ready to deal with this situation yet. There is a safe-
guarding WARN message when updating thread-local task statistics,
and it was applied in this situation.

This commit provides a workaround: we no longer try to update these
low-level stats for async update tasks. This means that the warning is
gone. But the statistics are (still) not available.

This resolves MID-7464.