Stabilize AbstractSynchronizationStoryTest
These tests were unstable from the day one. The concurrent nature
of synchronization tasks led to a whole spectrum of various failures.
But we don't consider them important, because we want to concentrate
only on the basic functionality of multi-resource synchronization here.

So now we changed these tests so that concurrent execution is avoided.

Other changes:

1. Some test artifacts (tasks, sync object template) were moved from
`common` directory to test-specific `sync-story` one. This is to lessen
the potential for coupling between the tests. (They weren't used in
other tests, anyway.)

2. The test code was modernized a bit.