Stop distribution of non-distributable activities
This commit introduced safety checks for activities that are known
to fail in multi-workers or multithreaded mode:

- report data aggregation (any parallelism is forbidden),
- any composite activity (multi-workers are forbidden, see the source).

This is an immediate fix for MID-7861.

However, it's not a complete one: we should check for multi-node
distributability _before_ the worker tasks are created, not in the tasks
themselves. Otherwise, the ultimate harm is avoided, but the workers
are already there - and suspended with a fatal error.