Add task suspension check for distributed reports
Now we skip writing per-bucket report data if the data-creation task
is going down. This is to ensure that the resulting data are clean.

In data-aggregation activity we now also check the sequence of partial
report data objects, to ensure nothing is missing and nothing is
included twice.

Notes:

1. There is (still) no check for the total expected number of partial
data objects in data aggregation activity. We rely on the activity
framework to ensure the aggregator is started only after the preceding
activity (data creation) is 100% complete.

2. In general, one should not resume report-creation activity after
it has been suspended (at least not after too much time). There is non-
negligible chance that the resulting report will be inconsistent because
of the time gap.

Related to MID-7839 (suspension of report tasks).