Re-implement midpoint.getFocusesByCorrelationRule
This method was not quite compatible with correlators changes introduced
in 4.5, and is definitely not compatible with synchronization schema
changes done in this release. So, it was rewritten from scratch, and
renamed to findCandidateOwners. (The old signature is still available
but marked as deprecated.) There may be minor differences in the
behavior, so please review javadoc and adapt your code, if necessary.

Related changes:

- added full focus list to CorrelationResult,
- generalized pre-inbounds processing to work without
a SynchronizationContext (i.e. within this method),
- some code simplifications in SynchronizationContext.

New (maybe) helpful class:

- ObjectSet, an OID-keyed object collection (TODO: better name)