Clean up synchronization-related code
This is a preparation for giant leaps in the area of resource
configuration changes planned in weeks to come.

1. Separated shadow tag generation from the shadow classification
(just to be precise in the use of terms).

2. Made further fields SynchronizationContext final. This makes
the data flow during context loading much cleaner.

3. Migrate from the use of kind/intent to use of object type definition.
This may be too brave, but hopefully reasonable. We don't want to deal
with messy combinations of account/unknown and similar classifications.

4. ResourceObjectClassifierImpl is now stand-alone component, not using
SynchronizationContext anymore.

5. Introduced lightweight ResourceObjectProcessingContext to be used
throughout synchronization-related code.

6. Added model expression environment for tag generation.

Work in progress. Some tests may fail.