Skip assignment approvals on focus deletion
When a focus is going to be deleted, assignment-level approvals
are simply skipped. There's no point in approving these changes,
as it is not clear what should be done in the case of their
rejection. In the need of focus deletion approvals, please use
appropriate policy constraints (referring to object deletion).

Resolves MID-7912.

(cherry picked from commit 43e0eb54de4e4688f5a3f6967067823369084777)