Add Correlator#checkCandidateOwner method
The recent rewrite of matchUserCorrelationRule that used the standard
correlation process was too simplistic: it failed in cases when the
candidate owner was not in repo yet. (Fortunately, our test suite
contains a test for this case!)

No simple solutions here - the special "checkCandidateOwner" method that
solves this case had to be conceived.

This resolves failing TestConsistencyMechanism in story tests.

Limitations: Not supported for more exotic correlators (expression,
composite) yet. Appropriate exception message informing the user to
turn off opportunistic synchronization is provided.