Improve resource expansion algorithm
Because we need to have definitions for all merged items, now we do the
expansion operation in two passes. First, we gather all the connector
configurations - to obtain the schema(s). The second pass will merge
the configurations with the appropriate definitions already applied.

Other changes:

1. Created ConnectorWithSchema to keep the schema along with the
connector object in a cleaner way. (And using this in ConnectorManager.)

2. ResourceManager#getConnectorSpec now throws ConfigurationException
(instead of SchemaExceptions) if it sees a misconfiguration like
a connector without a name or a reference.