Switch capabilities to static containers
1. CapabilityType is now a Containerable (was a simple bean only).

2. CapabilityCollectionType has been changed from "xsd:any" collection
of values to a statically-defined container. Related to this, it was
moved from common-3 to capabilities-3 namespace (to be close
to definition of individual capabilities).

All of this is related to implementation of resource and object type
inheritance - to be able to reasonably merge configured capabilities.
But this change (hopefully) simplifies any code that works with
capabilities.

Other changes:
- Naming of utility methods was fixed. For example, we no longer refer
  to "effective capabilities", as it may falsely indicate that such
  capabilities are enabled (they are not). Instances of
  "getEffectiveCapability" that returns enabled ones were renamed
  to "getEnabledCapability".
- Fixed a minor bug when updating connector capabilities metadata
  (wrong item path).

Work in progress. Some tests may fail.