Adapt test resources to recent changes
INCOMPATIBLE CHANGE (to release notes):

If you explicitly mention CapabilityCollectionType in your configuration
(typically as xsi:type="c:CapabilityCollectionType"), you have to change
this to "cap:CapabilityCollectionType".

This change was introduced in c91029fb651cea808aa72658900d4533de152686.
Here we just adapt a couple of test resources that refer to the legacy
name.